+++ Docs/Changes.md
@@ -15,6 +15,7 @@
   * Add: _Allow multiple type annotations for a single parameter_ ([Issue #618](https://github.com/vorov2/dyalect/issues/618), [docs](https://github.com/vorov2/dyalect/wiki/Functions#anno)).
   * Fix: _Correct errors for undefined types_ ([Issue #610](https://github.com/vorov2/dyalect/issues/610)).
   * Fix: _Test runner doesn't display file name in "onlyfailed" display mode_ ([Issue #614](https://github.com/vorov2/dyalect/issues/614)).
+  * Fix: _`Dictionary.ToLiteral` is not implemented correctly_ ([Issue #623](https://github.com/vorov2/dyalect/issues/623)).
 
 # 0.31.2
   * _Optimizations and cleanups in Dyalect console and Dyalect parser._